* commit 'af66f30bebdaf798c4394e4b33aeba952ff26522':
  Return std::size_t from member functions named bytes()
  Add member function to return size of the data type in bytes

+++ src/atlas/field/Field.h
@@ -149,7 +149,7 @@ public:
     size_t rank() const;
 
     /// @brief Number of bytes occupied by the values of this field
-    double bytes() const;
+    size_t bytes() const;
 
     /// @brief Output information of field
     friend std::ostream& operator<<( std::ostream& os, const Field& v );
